pyproject.toml
[project]
name = "sw-projection-storage"
version = "0.1.0"
description = "Spectrum projection data storage"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-cosmos>=4.9.0",
    "azure-storage-blob>=12.25.1",
    "fastapi[standard]>=0.115.13",
    "pydantic-settings>=2.10.1",
    "uvicorn>=0.34.3",
]

[dependency-groups]
dev = [
    "basedpyright>=1.29.5",
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "pytest-error-for-skips>=2.0.2",
]

[tool.pytest.ini_options]
pythonpath = "."

[tool.ruff.lint]
extend-select = ["E", "W"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["E722"]
